feat: new problem class system
In this PR, we added a customizable proclass for users, which they can
set to be publicly shared for others to use.
We improved the original proclass selection menu and added a new
interface for this purpose.
We also introduced a feature to collect proclass, allowing users to
gather their favorite proclass.