Misc: E2E test, migration, upgrade script, testcase setup UI, custom problem-class (#94)
* fix: missing third-party js script

* chore: add redis-db and service port install option

* refactor: we don't need multi-threading to do cleanup work

* refactor: there should be no distinction between kernels and users
Follow TOJ Spec

* feat(UI): feat: add password eye
Copy From TNFSH-Scoreboard

* feat: add db and redis migration

* feat: add simple upgrade script

* feat: problem add allow submit option

* feat: user add custom motto

* perf: incremental refresh challenge_state
The challenge_state is used to cache all challenge results.
When refreshing the challenge_state, the database will recalculate all data about the challenges, but many challenges are unnecessarily updated.
So the incremental refresh only updates challenges that have changed.

* refactor: remove unused problem expire field

* refactor: move get problem state from `list_pro()` to `map_acct_rate()`

* feat: redirect to sign-in page when user is a guest

* test: add e2e test

* feat: prevent unauthorized user from updating password

* test: add user motto test

* test: add allow_submit option test

* fix: change progress bar text

* fix: pdf file is displayed as garbled text

* feat: add a hash to check file integrity during upload

* fix: escape characters when content contain code block

* fix: do not push empty url to history

* feat: add testcase setup ui
In this PR, we add the full file management UI. Now we can use the web UI to manage our test case files, attachments, and checkers. We no longer need to upload a problem package to overwrite files to achieve this goal.
We also introduce multiple language limit settings, allowing us to set specific languages. This is important for Python 3 or Java because they are usually slower than compiled languages like C or C++.
We will provide multiple test case file operations, which can reduce many single and duplicate operations for users.

* feat: add a simple log viewer for log params

* feat: new problem class system

In this PR, we added a customizable proclass for users, which they can
set to be publicly shared for others to use.
We improved the original proclass selection menu and added a new
interface for this purpose.
We also introduced a feature to collect proclass, allowing users to
gather their favorite proclass.

* fix: follow the target behavior when a link has a target attribute

* refactor: move `self.acct` to the default namespace to reduce argument passing

* perf: use batch inserts to reduce SQL execution

* fix: corrected wrong acct_id in a multi-threading environment

* perf: use SQL to calculate user rank instead of Python
In TOJ, response time decreased from 2000ms to about 300ms

* fix: only update the code hash when the code is actually submitted

* ci: bypass installation restrictions of the package manager

* ci: remove unnecessary zip compression, as actions/upload-artifact will handle this

* fix: module not found error

* refactor: make ruff and pyright happy

* Refactor: Improve migration logic and error handling in main function

* fix: correct wrong log message grammar

* chore: remove `f.close()`, as the with statement handles this

---------

Co-authored-by: lifeadventurer <108756201+LifeAdventurer@users.noreply.github.com>