perf(test): remove AccountContext forces a one-second wait
Because tornado.create_signed_value uses int(clock()) as the timestamp,
it causes the signed values to be identical within the same second. This
leads to errors in AccountContext session operations. To avoid this,
AccountContext forces a one-second wait to ensure the timestamps are
different, which makes the test execution slower.

To address this, we introduced the dowhen package and changed the
timestamp from int(clock()) to clock(), so that AccountContext no longer
needs to wait for a second, thus speeding up the test process.