feat(contests): add system test similar to CF
System tests allow partial test data to be executed first in a contest to quickly verify results,
with the complete test data being executed after the contest ends.
This reduces the load on the judge system.
To support this, tags have been introduced for subtasks and test data to mark whether they are used for system tests.